Enemy Spawner upgrade, Coins despawn
Added a spawn queue to the enemy spawner script so enemies will get added to the queue if they try to spawn in the doorway when enemies are already there. Randomized the times a wave of enemies can spawn. Lowered the drop rate of coins and they can despawn. Coins blink when they get closer to despawning.